requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
aiosqlite
pyjwt
passlib[bcrypt]